# diasca-node/app/routers/sites.py
from typing import List
from uuid import UUID
from fastapi import APIRouter, Depends, HTTPException, status
from sqlalchemy.ext.asyncio import AsyncSession
from sqlmodel import select

from app.database import get_session
from app.models.site import Site
from app.schemas.operations import AttachGeoIdRequest, UpdateGeometryRequest, ValidateGeometryRequest
from app.schemas.common import PaginatedResponse

router = APIRouter(prefix="/sites", tags=["Sites"])

@router.get("/", response_model=PaginatedResponse[Site])
async def list_sites(
    offset: int = 0,
    limit: int = 100,
    session: AsyncSession = Depends(get_session)
):
    # This would eventually use actual auth and scopes
    result = await session.execute(select(Site).offset(offset).limit(limit))
    sites = result.scalars().all()
    # Mocking total count for scaffolding
    return PaginatedResponse(items=sites, total=len(sites), limit=limit, offset=offset)

@router.post("/", response_model=Site, status_code=status.HTTP_201_CREATED)
async def create_site(
    # In a real app, use a SiteCreate schema here, not the model directly
    site_data: dict,
    session: AsyncSession = Depends(get_session)
):
    # Mock endpoint
    pass

@router.get("/{site_id}", response_model=Site)
async def get_site(
    site_id: UUID,
    session: AsyncSession = Depends(get_session)
):
    site = await session.get(Site, site_id)
    if not site:
        raise HTTPException(status_code=404, detail="Site not found")
    return site

# --- Domain Operations ---

@router.post("/{site_id}:attachGeoId")
async def attach_geo_id(
    site_id: UUID,
    request: AttachGeoIdRequest,
    session: AsyncSession = Depends(get_session)
):
    # Mock endpoint for domain operation
    return {"status": "success", "site_id": site_id, "operation": "attachGeoId"}

@router.post("/{site_id}:updateGeometry")
async def update_geometry(
    site_id: UUID,
    request: UpdateGeometryRequest,
    session: AsyncSession = Depends(get_session)
):
    # Mock endpoint for domain operation
    return {"status": "success", "site_id": site_id, "operation": "updateGeometry"}

@router.post("/{site_id}:validateGeometry")
async def validate_geometry(
    site_id: UUID,
    request: ValidateGeometryRequest,
    session: AsyncSession = Depends(get_session)
):
    # Mock endpoint for domain operation
    return {"status": "success", "site_id": site_id, "operation": "validateGeometry", "is_valid": True}